move examples to mdx
