+++ packages/ui-extensions/docs/surfaces/point-of-sale/mdxExamples/discount-example/tile.ts
@@ -4,6 +4,8 @@ export default extension('pos.home.tile.render', (root, api) => {
   const shouldEnable = (subtotal: string): boolean => {
     return Number(subtotal) > 100;
   };
+
+  // [START tile.enable]
   // You can use the initial cart value to set up state
   const tile = root.createComponent(Tile, {
     title: 'Discount Example App',
@@ -11,11 +13,13 @@ export default extension('pos.home.tile.render', (root, api) => {
     enabled: shouldEnable(api.cart.subscribable.initial.subtotal),
     onPress: api.action.presentModal,
   });
+  // [END tile.enable]
 
+  // [START tile.subscribe]
   // You can subscribe to changes in the cart to mutate state
   api.cart.subscribable.subscribe((cart) => {
     tile.updateProps({enabled: shouldEnable(cart.subtotal)});
   });
-
+  // [END tile.subscribe]
   root.append(tile);
 });
