+++ packages/ui-extensions/docs/surfaces/point-of-sale/mdxExamples/print-example/print-tile.tsx
