--- packages/ui-extensions/docs/surfaces/point-of-sale/staticPages/examples/discount-example/1-enable-the-tile.ts
@@ -1,6 +1,0 @@
-const tile = root.createComponent(Tile, {
-  title: 'Discount Example App',
-  subtitle: 'Javascript',
-  enabled: shouldEnable(api.cart.subscribable.initial.subtotal),
-  onPress: api.action.presentModal
-});
