--- packages/ui-extensions/docs/surfaces/point-of-sale/staticPages/examples/discount-example/1-enable-the-tile.tsx
@@ -1,1 +1,0 @@
-const [enabled, setEnabled] = useState(shouldEnable(api.cart.subscribable.initial.subtotal));
