--- packages/ui-extensions/docs/surfaces/point-of-sale/staticPages/examples/discount-example/2-subscribe-to-cart.tsx
@@ -1,3 +1,0 @@
-api.cart.subscribable.subscribe((cart) => {
-  setEnabled(shouldEnable(cart.subtotal));
-});
