--- packages/ui-extensions/docs/surfaces/point-of-sale/staticPages/examples/discount-example/3-add-buttons.ts
@@ -1,12 +1,0 @@
-const percentageDiscountButton = root.createComponent(Button, {
-  title: '25%',
-  onPress: () => { onButtonPress('Percentage', '25% off', '25') }
-});
-
-const fixedAmountDiscountButton = root.createComponent(Button, {
-  title: '$10',
-  onPress: () => { onButtonPress('FixedAmount', '$10 off', '10') }
-});
-
-scrollView.appendChild(percentageDiscountButton);
-scrollView.appendChild(fixedAmountDiscountButton);
