--- packages/ui-extensions/docs/surfaces/point-of-sale/staticPages/examples/discount-example/3-add-buttons.tsx
@@ -1,4 +1,0 @@
-<ScrollView>
-  <Button title='25%' onPress={() => onButtonPress('Percentage', '25% off', '25')} />
-  <Button title='$10' onPress={() => onButtonPress('FixedAmount', '$10 off', '10')} />
-</ScrollView>
