--- packages/ui-extensions/docs/surfaces/point-of-sale/staticPages/examples/discount-example/4-define-onpress.ts
@@ -1,4 +1,0 @@
-const onButtonPress = (type: DiscountType, title: string, amount: string) => {
-  api.cart.applyCartDiscount(type, title, amount);
-  api.toast.show('Discount applied');
-};
