--- packages/ui-extensions/docs/surfaces/point-of-sale/staticPages/examples/discount-example/4-define-onpress.tsx
@@ -1,1 +1,0 @@
-onPress={() => onButtonPress('FixedAmount', '$10 off', '10')}
